set event feature image which event has no image

+++ includes/class-import-facebook-events-tec.php
@@ -141,11 +141,11 @@ class Import_Facebook_Events_TEC {
 		global $ife_events;
 
 		$is_exitsing_event = $ife_events->common->get_event_by_event_id( $this->event_posttype, $centralize_array['ID'] );
+		$options           = ife_get_import_options( $centralize_array['origin'] );
 		$formated_args = array();
         
 		if ( $is_exitsing_event && is_numeric( $is_exitsing_event ) && $is_exitsing_event > 0 ) {
 
-			$options       = ife_get_import_options( $centralize_array['origin'] );
 			$update_events = isset( $options['update_events'] ) ? $options['update_events'] : 'no';
 			$skip_trash    = isset( $options['skip_trash'] ) ? $options['skip_trash'] : 'no';
 			$post_status   = get_post_status( $is_exitsing_event );
@@ -254,6 +254,11 @@ class Import_Facebook_Events_TEC {
 			$event_featured_image = $centralize_array['image_url'];
 			if ( ! empty( $event_featured_image ) ) {
 				$ife_events->common->setup_featured_image_to_event( $new_event_id, $event_featured_image );
+			}else{
+				$default_thumb  = isset( $options['ife_event_default_thumbnail'] ) ? $options['ife_event_default_thumbnail'] : '';
+				if( !empty( $default_thumb ) ){
+					set_post_thumbnail( $new_event_id, $default_thumb );
+				}
 			}
 
 			//Insert in Custom Table 
@@ -371,8 +376,13 @@ class Import_Facebook_Events_TEC {
 			$event_featured_image = $centralize_array['image_url'];
 			if ( ! empty( $event_featured_image ) ) {
 				$ife_events->common->setup_featured_image_to_event( $update_event_id, $event_featured_image );
-			} else {
-				delete_post_thumbnail( $update_event_id );
+			}else{
+				$default_thumb  = isset( $options['ife_event_default_thumbnail'] ) ? $options['ife_event_default_thumbnail'] : '';
+				if( !empty( $default_thumb ) ){
+					set_post_thumbnail( $update_event_id, $default_thumb );
+				}else{
+					delete_post_thumbnail( $update_event_id );
+				}
 			}
